meta/requirements.txt
openai>=1.30.1
PyYAML>=6.0
aiofiles>=23.2.1
rich>=13.7.0
tqdm>=4.66.2
pathspec>=0.11.1
python-dotenv>=1.1.0